+++ ui/web/public/ui-web.generated.d.ts
@@ -16,11 +16,12 @@ declare module '@yuants/ui-web' {
     RegistryWidgetsType,
   } from '@rjsf/utils';
   import { ButtonProps } from '@douyinfe/semi-ui/lib/es/button';
+  import { ToastReactProps } from '@douyinfe/semi-ui/lib/es/toast';
   import * as _supabase_supabase_js from '@supabase/supabase-js';
   import { User } from '@supabase/supabase-js';
   import { Terminal } from '@yuants/protocol';
 
-  namespace index_d$w {
+  namespace index_d$x {
     export {};
   }
 
@@ -36,30 +37,24 @@ declare module '@yuants/ui-web' {
 
   const useAccountInfo: (account_id: string) => rxjs.Observable<_yuants_data_model.IAccountInfo>;
 
-  const index_d$v_AccountSelector: typeof AccountSelector;
-  const index_d$v_InlineAccountId: typeof InlineAccountId;
-  const index_d$v_useAccountInfo: typeof useAccountInfo;
-  namespace index_d$v {
+  const index_d$w_AccountSelector: typeof AccountSelector;
+  const index_d$w_InlineAccountId: typeof InlineAccountId;
+  const index_d$w_useAccountInfo: typeof useAccountInfo;
+  namespace index_d$w {
     export {
-      index_d$v_AccountSelector as AccountSelector,
-      index_d$v_InlineAccountId as InlineAccountId,
-      index_d$v_useAccountInfo as useAccountInfo,
+      index_d$w_AccountSelector as AccountSelector,
+      index_d$w_InlineAccountId as InlineAccountId,
+      index_d$w_useAccountInfo as useAccountInfo,
     };
+  }
+
+  namespace index_d$v {
+    export {};
   }
 
   namespace index_d$u {
     export {};
   }
-
-  namespace index_d$t {
-    export {};
-  }
-
-  /**
-   * BIOS Status Component
-   * @public
-   */
-  const BIOS: React.MemoExoticComponent<() => react_jsx_runtime.JSX.Element>;
 
   const createPersistBehaviorSubject: <T>(key: string, initialValue: T) => BehaviorSubject<T | undefined>;
 
@@ -68,19 +63,20 @@ declare module '@yuants/ui-web' {
    * @public
    */
   const ready$: ReplaySubject<unknown>;
+  const error$: ReplaySubject<unknown>;
 
-  const index_d$s_BIOS: typeof BIOS;
-  const index_d$s_createPersistBehaviorSubject: typeof createPersistBehaviorSubject;
-  const index_d$s_ready$: typeof ready$;
-  namespace index_d$s {
+  const index_d$t_createPersistBehaviorSubject: typeof createPersistBehaviorSubject;
+  const index_d$t_error$: typeof error$;
+  const index_d$t_ready$: typeof ready$;
+  namespace index_d$t {
     export {
-      index_d$s_BIOS as BIOS,
-      index_d$s_createPersistBehaviorSubject as createPersistBehaviorSubject,
-      index_d$s_ready$ as ready$,
+      index_d$t_createPersistBehaviorSubject as createPersistBehaviorSubject,
+      index_d$t_error$ as error$,
+      index_d$t_ready$ as ready$,
     };
   }
 
-  namespace index_d$r {
+  namespace index_d$s {
     export {};
   }
 
@@ -88,30 +84,30 @@ declare module '@yuants/ui-web' {
   const executeCommand: (id: string, params?: {}) => Promise<void>;
   const CommandCenter: React.MemoExoticComponent<() => react_jsx_runtime.JSX.Element>;
 
-  const index_d$q_CommandCenter: typeof CommandCenter;
-  const index_d$q_executeCommand: typeof executeCommand;
-  const index_d$q_registerCommand: typeof registerCommand;
-  namespace index_d$q {
+  const index_d$r_CommandCenter: typeof CommandCenter;
+  const index_d$r_executeCommand: typeof executeCommand;
+  const index_d$r_registerCommand: typeof registerCommand;
+  namespace index_d$r {
     export {
-      index_d$q_CommandCenter as CommandCenter,
-      index_d$q_executeCommand as executeCommand,
-      index_d$q_registerCommand as registerCommand,
+      index_d$r_CommandCenter as CommandCenter,
+      index_d$r_executeCommand as executeCommand,
+      index_d$r_registerCommand as registerCommand,
     };
+  }
+
+  namespace index_d$q {
+    export {};
   }
 
   namespace index_d$p {
     export {};
   }
 
-  namespace index_d$o {
-    export {};
-  }
-
   const useValue: <T>(id: string, initialValue: T) => [T, (v: T) => void];
 
-  const index_d$n_useValue: typeof useValue;
-  namespace index_d$n {
-    export { index_d$n_useValue as useValue };
+  const index_d$o_useValue: typeof useValue;
+  namespace index_d$o {
+    export { index_d$o_useValue as useValue };
   }
 
   interface IDataRecordViewDef<T> {
@@ -132,23 +128,23 @@ declare module '@yuants/ui-web' {
    */
   function DataRecordView<T>(props: IDataRecordViewDef<T>): react_jsx_runtime.JSX.Element;
 
-  const index_d$m_DataRecordView: typeof DataRecordView;
-  namespace index_d$m {
-    export { index_d$m_DataRecordView as DataRecordView };
+  const index_d$n_DataRecordView: typeof DataRecordView;
+  namespace index_d$n {
+    export { index_d$n_DataRecordView as DataRecordView };
   }
 
-  namespace index_d$l {
+  namespace index_d$m {
     export {};
   }
 
   const DesktopLayout: () => react_jsx_runtime.JSX.Element | null;
 
-  const index_d$k_DesktopLayout: typeof DesktopLayout;
-  namespace index_d$k {
-    export { index_d$k_DesktopLayout as DesktopLayout };
+  const index_d$l_DesktopLayout: typeof DesktopLayout;
+  namespace index_d$l {
+    export { index_d$l_DesktopLayout as DesktopLayout };
   }
 
-  namespace index_d$j {
+  namespace index_d$k {
     export {};
   }
 
@@ -166,10 +162,10 @@ declare module '@yuants/ui-web' {
     version: string;
   }>;
 
-  const index_d$i_loadTgzBlob: typeof loadTgzBlob;
-  const index_d$i_resolveVersion: typeof resolveVersion;
-  namespace index_d$i {
-    export { index_d$i_loadTgzBlob as loadTgzBlob, index_d$i_resolveVersion as resolveVersion };
+  const index_d$j_loadTgzBlob: typeof loadTgzBlob;
+  const index_d$j_resolveVersion: typeof resolveVersion;
+  namespace index_d$j {
+    export { index_d$j_loadTgzBlob as loadTgzBlob, index_d$j_resolveVersion as resolveVersion };
   }
 
   interface IFileSystemStatResult {
@@ -241,22 +237,22 @@ declare module '@yuants/ui-web' {
   }
   function resolve(x: string, options: IOptions): Promise<string>;
 
-  const index_d$h_FsBackend$: typeof FsBackend$;
-  const index_d$h_bundleCode: typeof bundleCode;
-  const index_d$h_fs: typeof fs;
-  const index_d$h_historyWorkspaceRoot$: typeof historyWorkspaceRoot$;
-  const index_d$h_replaceWorkspaceRoot: typeof replaceWorkspaceRoot;
-  const index_d$h_resolve: typeof resolve;
-  const index_d$h_workspaceRoot$: typeof workspaceRoot$;
-  namespace index_d$h {
+  const index_d$i_FsBackend$: typeof FsBackend$;
+  const index_d$i_bundleCode: typeof bundleCode;
+  const index_d$i_fs: typeof fs;
+  const index_d$i_historyWorkspaceRoot$: typeof historyWorkspaceRoot$;
+  const index_d$i_replaceWorkspaceRoot: typeof replaceWorkspaceRoot;
+  const index_d$i_resolve: typeof resolve;
+  const index_d$i_workspaceRoot$: typeof workspaceRoot$;
+  namespace index_d$i {
     export {
-      index_d$h_FsBackend$ as FsBackend$,
-      index_d$h_bundleCode as bundleCode,
-      index_d$h_fs as fs,
-      index_d$h_historyWorkspaceRoot$ as historyWorkspaceRoot$,
-      index_d$h_replaceWorkspaceRoot as replaceWorkspaceRoot,
-      index_d$h_resolve as resolve,
-      index_d$h_workspaceRoot$ as workspaceRoot$,
+      index_d$i_FsBackend$ as FsBackend$,
+      index_d$i_bundleCode as bundleCode,
+      index_d$i_fs as fs,
+      index_d$i_historyWorkspaceRoot$ as historyWorkspaceRoot$,
+      index_d$i_replaceWorkspaceRoot as replaceWorkspaceRoot,
+      index_d$i_resolve as resolve,
+      index_d$i_workspaceRoot$ as workspaceRoot$,
     };
   }
 
@@ -297,25 +293,25 @@ declare module '@yuants/ui-web' {
    */
   const showForm: <T>(schema: JSONSchema7, initialData?: any) => Promise<T>;
 
-  const index_d$g_Form: typeof Form;
-  const index_d$g_Theme: typeof Theme;
-  const index_d$g_generateForm: typeof generateForm;
-  const index_d$g_generateTemplates: typeof generateTemplates;
-  const index_d$g_generateTheme: typeof generateTheme;
-  const index_d$g_generateWidgets: typeof generateWidgets;
-  const index_d$g_showForm: typeof showForm;
-  namespace index_d$g {
+  const index_d$h_Form: typeof Form;
+  const index_d$h_Theme: typeof Theme;
+  const index_d$h_generateForm: typeof generateForm;
+  const index_d$h_generateTemplates: typeof generateTemplates;
+  const index_d$h_generateTheme: typeof generateTheme;
+  const index_d$h_generateWidgets: typeof generateWidgets;
+  const index_d$h_showForm: typeof showForm;
+  namespace index_d$h {
     export {
-      index_d$g_Form as Form,
+      index_d$h_Form as Form,
       _default$1 as Templates,
-      index_d$g_Theme as Theme,
+      index_d$h_Theme as Theme,
       _default as Widgets,
       Form as default,
-      index_d$g_generateForm as generateForm,
-      index_d$g_generateTemplates as generateTemplates,
-      index_d$g_generateTheme as generateTheme,
-      index_d$g_generateWidgets as generateWidgets,
-      index_d$g_showForm as showForm,
+      index_d$h_generateForm as generateForm,
+      index_d$h_generateTemplates as generateTemplates,
+      index_d$h_generateTheme as generateTheme,
+      index_d$h_generateWidgets as generateWidgets,
+      index_d$h_showForm as showForm,
     };
   }
 
@@ -370,8 +366,13 @@ declare module '@yuants/ui-web' {
       total_profit: number;
     };
     investors: Record<string, InvestorMeta>;
+    investor_cashflow: Record<string, InvestorCashFlowItem[]>;
     investor_derived: Record<string, InvestorInfoDerived>;
     events: IFundEvent[];
+  }
+  interface InvestorCashFlowItem {
+    updated_at: number;
+    deposit: number;
   }
   interface InvestorMeta {
     /** 姓名 */
@@ -384,11 +385,17 @@ declare module '@yuants/ui-web' {
     deposit: number;
     /** 税率 */
     tax_rate: number;
+    /** 创建时间 */
+    created_at: number;
   }
   /**
    * 投资人信息的计算衍生数据
    */
   interface InvestorInfoDerived {
+    /** 持有时间 */
+    holding_days: number;
+    /** 资产在时间上的积分 */
+    timed_assets: number;
     /** 税前资产 */
     pre_tax_assets: number;
     /** 应税额 */
@@ -401,6 +408,7 @@ declare module '@yuants/ui-web' {
     after_tax_profit: number;
     /** 税后收益率 */
     after_tax_profit_rate: number;
+    after_tax_IRR: number;
     /** 税后份额 */
     after_tax_share: number;
     /** 份额占比 */
@@ -412,11 +420,11 @@ declare module '@yuants/ui-web' {
     }
   }
 
-  namespace index_d$f {
+  namespace index_d$g {
     export {};
   }
 
-  namespace index_d$e {
+  namespace index_d$f {
     export {};
   }
 
@@ -449,33 +457,81 @@ declare module '@yuants/ui-web' {
 
   function TableView<T>(props: { table: Table<T> }): react_jsx_runtime.JSX.Element;
 
-  const index_d$d_Button: typeof Button;
-  const index_d$d_DataView: typeof DataView;
-  type index_d$d_IPivotTableProps<T> = IPivotTableProps<T>;
-  const index_d$d_ListView: typeof ListView;
-  const index_d$d_PivotTable: typeof PivotTable;
-  const index_d$d_TableView: typeof TableView;
-  namespace index_d$d {
+  type ToastProps = string | Omit<ToastReactProps, 'type'>;
+  /**
+   * Yuan Toast Component
+   */
+  const Toast: {
+    info: (props: ToastProps) => string;
+    error: (props: ToastProps) => string;
+    success: (props: ToastProps) => string;
+    warning: (props: ToastProps) => string;
+    close: (id: string) => void;
+  };
+
+  const index_d$e_Button: typeof Button;
+  const index_d$e_DataView: typeof DataView;
+  type index_d$e_IPivotTableProps<T> = IPivotTableProps<T>;
+  const index_d$e_ListView: typeof ListView;
+  const index_d$e_PivotTable: typeof PivotTable;
+  const index_d$e_TableView: typeof TableView;
+  const index_d$e_Toast: typeof Toast;
+  type index_d$e_ToastProps = ToastProps;
+  namespace index_d$e {
     export {
-      index_d$d_Button as Button,
-      index_d$d_DataView as DataView,
-      type index_d$d_IPivotTableProps as IPivotTableProps,
-      index_d$d_ListView as ListView,
-      index_d$d_PivotTable as PivotTable,
-      index_d$d_TableView as TableView,
+      index_d$e_Button as Button,
+      index_d$e_DataView as DataView,
+      type index_d$e_IPivotTableProps as IPivotTableProps,
+      index_d$e_ListView as ListView,
+      index_d$e_PivotTable as PivotTable,
+      index_d$e_TableView as TableView,
+      index_d$e_Toast as Toast,
+      type index_d$e_ToastProps as ToastProps,
     };
   }
 
-  namespace index_d$c {
+  namespace index_d$d {
     export {};
+  }
+
+  /**
+   * @public
+   */
+  const Launch: React.MemoExoticComponent<() => react_jsx_runtime.JSX.Element>;
+
+  const index_d$c_Launch: typeof Launch;
+  namespace index_d$c {
+    export { index_d$c_Launch as Launch };
   }
 
   namespace index_d$b {
     export {};
   }
 
+  interface IInterleavingConfigItem {
+    account_id: string;
+    datasource_id: string;
+    product_id: string;
+    order_type: string;
+    order_direction: string;
+    volume: number;
+    disabled?: boolean;
+  }
+  interface IInterleavingConfig {
+    count: number;
+    items: IInterleavingConfigItem[];
+  }
+  const InterleavingTraderConfig$: BehaviorSubject<IInterleavingConfig | undefined>;
+
+  type index_d$a_IInterleavingConfig = IInterleavingConfig;
+  type index_d$a_IInterleavingConfigItem = IInterleavingConfigItem;
+  const index_d$a_InterleavingTraderConfig$: typeof InterleavingTraderConfig$;
   namespace index_d$a {
-    export {};
+    export {
+      type index_d$a_IInterleavingConfig as IInterleavingConfig,
+      type index_d$a_IInterleavingConfigItem as IInterleavingConfigItem,
+      index_d$a_InterleavingTraderConfig$ as InterleavingTraderConfig$,
+    };
   }
 
   const LocalizePageTitle: React.ComponentType<{
@@ -599,18 +655,13 @@ declare module '@yuants/ui-web' {
   const toggleShowHome: () => void;
   const HomePage: React.MemoExoticComponent<() => react_jsx_runtime.JSX.Element | null>;
 
-  const secretURL: (url: string) => string;
-  interface ICryptoHostConfig {
-    label: string;
-    public_key: string;
-    private_key: string;
-    host_url: string;
-  }
-  const cryptoHosts$: rxjs.BehaviorSubject<ICryptoHostConfig[] | undefined>;
-  const network$: rxjs.Observable<string[]>;
-
-  const isDarkMode$: rxjs.BehaviorSubject<boolean | undefined>;
+  const DarkModeSetting$: rxjs.BehaviorSubject<'auto' | 'light' | 'dark' | undefined>;
+  const isDarkMode$: Observable<boolean>;
   const useIsDarkMode: () => boolean;
+
+  const DarkModeEffect: () => react_jsx_runtime.JSX.Element;
+
+  const DarkmodeSwitch: React.MemoExoticComponent<() => react_jsx_runtime.JSX.Element>;
 
   /**
    * Hook to use the page closing confirm
@@ -622,6 +673,18 @@ declare module '@yuants/ui-web' {
    * for example, some component with complex form data that the user might lose if they close the page.
    */
   const usePageClosingConfirm: (disabled?: boolean) => void;
+
+  const FullScreenButton: () => react_jsx_runtime.JSX.Element;
+
+  const secretURL: (url: string) => string;
+  interface ICryptoHostConfig {
+    label: string;
+    public_key: string;
+    private_key: string;
+    host_url: string;
+  }
+  const cryptoHosts$: rxjs.BehaviorSubject<ICryptoHostConfig[] | undefined>;
+  const network$: rxjs.Observable<string[]>;
 
   interface IHostConfigItem {
     name: string;
@@ -635,6 +698,10 @@ declare module '@yuants/ui-web' {
   }>;
   const OHLCIdList$: BehaviorSubject<string[]>;
 
+  const index_d$1_DarkModeEffect: typeof DarkModeEffect;
+  const index_d$1_DarkModeSetting$: typeof DarkModeSetting$;
+  const index_d$1_DarkmodeSwitch: typeof DarkmodeSwitch;
+  const index_d$1_FullScreenButton: typeof FullScreenButton;
   const index_d$1_HomePage: typeof HomePage;
   type index_d$1_IHostConfigItem = IHostConfigItem;
   const index_d$1_OHLCIdList$: typeof OHLCIdList$;
@@ -651,6 +718,10 @@ declare module '@yuants/ui-web' {
   const index_d$1_usePageClosingConfirm: typeof usePageClosingConfirm;
   namespace index_d$1 {
     export {
+      index_d$1_DarkModeEffect as DarkModeEffect,
+      index_d$1_DarkModeSetting$ as DarkModeSetting$,
+      index_d$1_DarkmodeSwitch as DarkmodeSwitch,
+      index_d$1_FullScreenButton as FullScreenButton,
       index_d$1_HomePage as HomePage,
       type index_d$1_IHostConfigItem as IHostConfigItem,
       index_d$1_OHLCIdList$ as OHLCIdList$,
@@ -691,27 +762,28 @@ declare module '@yuants/ui-web' {
   }
 
   export {
-    index_d$w as AccountComposition,
-    index_d$v as AccountInfo,
-    index_d$u as AccountRiskInfo,
-    index_d$t as Agent,
-    index_d$s as BIOS,
-    index_d$r as Chart,
-    index_d$q as CommandCenter,
-    index_d$p as Copilot,
-    index_d$o as CopyDataRelation,
-    index_d$n as Data,
-    index_d$m as DataRecord,
-    index_d$l as Deploy,
-    index_d$k as DesktopLayout,
-    index_d$j as Editor,
-    index_d$i as Extensions,
-    index_d$h as FileSystem,
-    index_d$g as Form,
-    index_d$f as Fund,
-    index_d$e as GeneralSpecificRelations,
-    index_d$d as Interactive,
-    index_d$c as Kernel,
+    index_d$x as AccountComposition,
+    index_d$w as AccountInfo,
+    index_d$v as AccountRiskInfo,
+    index_d$u as Agent,
+    index_d$t as BIOS,
+    index_d$s as Chart,
+    index_d$r as CommandCenter,
+    index_d$q as Copilot,
+    index_d$p as CopyDataRelation,
+    index_d$o as Data,
+    index_d$n as DataRecord,
+    index_d$m as Deploy,
+    index_d$l as DesktopLayout,
+    index_d$k as Editor,
+    index_d$j as Extensions,
+    index_d$i as FileSystem,
+    index_d$h as Form,
+    index_d$g as Fund,
+    index_d$f as GeneralSpecificRelations,
+    index_d$e as Interactive,
+    index_d$d as Kernel,
+    index_d$c as Launch,
     index_d$b as Market,
     index_d$a as Order,
     index_d$9 as Pages,